Extracted regex::Class and related types to class module.
